feat: update .js to .ts
