+++ .babelrc.js
@@ -1,0 +1,13 @@
+module.exports = {
+  presets: [
+    '@babel/preset-typescript',
+    [
+      '@babel/preset-env',
+      {
+        targets: {
+          node: '12'
+        }
+      }
+    ]
+  ]
+};
