+++ .eslintignore
@@ -1,0 +1,3 @@
+lib/
+renovate.json
+tsconfig.json
