+++ .gitignore
@@ -1,6 +1,7 @@
 .idea/
 coverage/
 node_modules/
+lib/
 temp
 yarn.lock
 *.log
