+++ tsconfig.json
@@ -1,0 +1,13 @@
+{
+  "compilerOptions": {
+    "allowSyntheticDefaultImports": true,
+    "esModuleInterop": true,
+    "isolatedModules": true,
+    "module": "commonjs",
+    "skipLibCheck": true,
+    "sourceMap": false,
+    "target": "esnext"
+  },
+  "exclude": ["node_modules"],
+  "include": ["src"]
+}
